Add styling to all files
Added 2 files for styling and updated the structure of the HTML.
Added a folder for the CSS with a file for the index and recipes.
Adjusted the HTML of all files so the css would work across multiple
files by following a similar structure for all recipe pages. styling
is consistent across the board.